ZON-5694: Tweak XML config format, implement image
