+++ core/src/zeit/content/modules/tests/fixtures/newsletter.xml
@@ -1,0 +1,8 @@
+<newsletters>
+  <newsletter id="zplus">
+    <title>Z+/Abonnenten</title>
+    <image>http://xml.zeit.de/test/wosc/testbild/</image>
+    <text>Hier anmelden</text>
+    <text_anonymous>Registrieren und dann anmelden</text_anonymous>
+  </newsletter>
+</newsletters>
